[src/log_plotter/datalogger_plotter_with_pyqtgraph.py, log_parser.py, plot_utils.py] Separate file
