+++ src/log_plotter/datalogger_plotter_with_pyqtgraph.py
@@ -1,25 +1,18 @@
 #!/usr/bin/env python
 
-import argparse
-import csv
 import functools
+import argparse
 import math
-import multiprocessing
-import numpy
 import sys
-import time
 import yaml
-import metayaml
-import os
-import fnmatch
-import re
 import signal
-import log_plotter.plot_method as plot_method
-from log_plotter.graph_legend import GraphLegendInfo, expand_str_to_list
-import log_plotter.yaml_selector as yaml_selector
 import log_plotter.pyqtgraph_LegendItem_patch
+import log_plotter.plot_method as plot_method
+import log_plotter.yaml_selector as yaml_selector
 import log_plotter.graph_tools as graph_tools
-
+from log_plotter.graph_legend import GraphLegendInfo
+from log_plotter.plot_utils import my_time
+from log_plotter.log_parser import LogParser
 try:
     import pyqtgraph
 except:
@@ -27,123 +20,40 @@ except:
     sys.exit(1)
 
 
-# decorator for time measurement
-def my_time(func):
-    @functools.wraps(func)
-    def wrapper(*args, **kwargs):
-        start = time.time()
-        print("start : " + func.func_name)
-        result = func(*args, **kwargs)
-        print('{0:>10} : {1:3.3f} [s]'.format(func.func_name,
-                                              time.time() - start))
-        return result
-    return wrapper
-
-
-# seems that we should declare global function for multiprocess
-def readOneTopic(fname):
-    tmp = []
-    try:
-        with open(fname, 'r') as f:
-            reader = csv.reader(f, delimiter=' ')
-            for row in reader:
-                dl = filter(lambda x: x != '', row)
-                tmp.append([float(x) for x in dl])
-    except Exception as e:
-        print '[readOneToopic] error occured while reading {}'.format(fname)
-        raise e
-    return numpy.array(tmp)
-
-# return file list matching specific substring
-def findFile(pattern, path):
-    result = []
-    for root, dirs, files in os.walk(path):
-        for name in files:
-            if fnmatch.fnmatch(name, pattern):
-                result.append(os.path.join(root, name))
-    return result
-
-# replace RobotHardware0 to each simulation one
-def replaceRH(fname_list):
-    log_dir = os.path.dirname(os.path.abspath(fname_list[0]))
-    base_name = os.path.splitext(os.path.basename(fname_list[0]))[0]
-    # choreonoid
-    if findFile(base_name + '.RobotHardware_choreonoid0_*', log_dir):
-        fname_list = [fname.replace('.RobotHardware0_', '.RobotHardware_choreonoid0_') for fname in fname_list]
-        return fname_list
-    # hrpsys simulator
-    elif findFile(base_name + '.*(Robot)0_*', log_dir):
-        # pick up robot name (.<ROBOT_NAME>(Robot)0_)
-        rep = re.sub(findFile(base_name + '.*(Robot)0_*', log_dir)[0].split('(Robot)0_')[1], '',
-                     os.path.splitext(findFile(base_name + '.*(Robot)0_*', log_dir)[0])[1])
-        fname_list = [fname.replace('.RobotHardware0_', rep) for fname in fname_list]
-        return fname_list
-    else: return fname_list
-
-class DataloggerLogParser:
+class LogPlotter(object):
     def __init__(self, fname, plot_conf_name, layout_conf_name, title):
+        '''
+        :param str fname: file name of log
+        :param str plot_conf_name: plot yaml file name
+        :param str layout_conf_name: layout yaml file name
+        :param str title: window title
+        '''
+        # set args
         self.fname = fname
-        self.plot_dict = metayaml.read(plot_conf_name)
-        self.layout_dict = metayaml.read(layout_conf_name)["main"]
-        # expand [0-33] => [0,1,2,...,33]
-        for leg_info in self.plot_dict.values():
-            for log_info in leg_info['data']:
-                if type(log_info['column'][0]) == str:
-                    log_info['column'] = expand_str_to_list(log_info['column'][0])
-        for dict_title in self.layout_dict:
-            for leg in self.layout_dict[dict_title]['legends']:
-                if type(leg['id'][0]) == str:
-                    leg['id'] = expand_str_to_list(leg['id'][0])
-            self.layout_dict[dict_title].setdefault('newline', True)
-            self.layout_dict[dict_title].setdefault('left_label', False)
-            self.layout_dict[dict_title].setdefault('bottom_label', "time [s]")
-
+        self.plot_conf_name = plot_conf_name
+        self.layout_conf_name = layout_conf_name
         # setup view
         self.view = pyqtgraph.GraphicsLayoutWidget()
         self.view.setBackground('w')
         self.view.setWindowTitle(title if title else fname.split('/')[-1])
-        # self.dateListDict is set by self.readData()
-        self.dataListDict = {}# todo: to list of dictionary
         # back up for plot items
         self.plotItemOrig = {}
-
         # default font style
         self.font_type = 'Times New Roman'
         self.font_size = 12
         self.font_color = 'black'
 
     @my_time
-    def readData(self):
+    def getData(self):
         '''
-        read log data from log files and store dataListDict
-        # self.dataListDict[topic] = numpy.array([[t_0, x_0, y_0, ...],
-        #                                         [t_1, x_1, y_1, ...],
-        #                                         ...,
-        #                                         [t_n, x_n, y_n, ...]])
+        get data using LogParser
         '''
-        # get list fo used topic
-        all_legends = reduce(lambda x,y: x+y, [self.layout_dict[title]['legends'] for title in self.layout_dict])
-        used_keys = list(set([legend['key'] for legend in all_legends]))
-        log_col_pairs = reduce(lambda x,y: x+y, [self.plot_dict[key]['data'] for key in used_keys])
-        topic_list = list(set([log_col_pair['log'] for log_col_pair in log_col_pairs]))
-        self._used_keys = used_keys
-        self._topic_list = topic_list
-
-        # store data in parallel
-        fname_list = replaceRH([self.fname+'.'+ext for ext in topic_list])
-        pl = multiprocessing.Pool()
-        data_list = pl.map(readOneTopic, fname_list)
-        for topic, data in zip(topic_list, data_list):
-            self.dataListDict[topic] = data
-        # set the fastest time as 0
-        min_time = min([self.dataListDict[topic][0][0] for topic in topic_list])
-        for topic in topic_list:
-            raw_time = self.dataListDict[topic][:, 0]
-            self.dataListDict[topic][:, 0] = [x - min_time for x in raw_time]
-        # fix servoState
-        if 'RobotHardware0_servoState' in topic_list:
-            ss_tmp = self.dataListDict['RobotHardware0_servoState'][:, 1:]
-            self.dataListDict['RobotHardware0_servoState'][:, 1:] = numpy.fromstring(ss_tmp.astype('i').tostring(), dtype='f').reshape(ss_tmp.shape)
+        log_parser = LogParser(self.fname, self.plot_conf_name, self.layout_conf_name)
+        log_parser.readData()
+        self.plot_dict = log_parser.plot_dict
+        self.layout_dict = log_parser.layout_dict
+        self._topic_list = log_parser._topic_list
+        self.dataListDict = log_parser.dataListDict
 
     @my_time
     def setLayout(self):
@@ -419,7 +329,7 @@ class DataloggerLogParser:
 
     def main(self):
         '''
-        1. read log files
+        1. get data
         2. decide layout
         3. plot data
         4. set label
@@ -427,7 +337,7 @@ class DataloggerLogParser:
         6. customize context menu
         7. show
         '''
-        self.readData()
+        self.getData()
         self.setLayout()
         self.linkAxes()
         self.plotData()
@@ -453,7 +363,7 @@ def main():
     if args.plot is None or args.layout is None: # check args
         get_yamls_path = yaml_selector.MainDialog()
         args.plot, args.layout = get_yamls_path()
-    a = DataloggerLogParser(args.f, args.plot, args.layout, args.t)
+    a = LogPlotter(args.f, args.plot, args.layout, args.t)
     a.main()
 
     if args.i:
@@ -461,7 +371,7 @@ def main():
         # start ipython
         print '====='
         print "please use \033[33mapp.processEvents()\033[m to update graph."
-        print "you can use \033[33ma\033[m as DataloggerLogParser instance."
+        print "you can use \033[33ma\033[m as LogPlotter instance."
         print '====='
         from IPython import embed
         embed()
